--- bloom-inference-server/ds_inference/__init__.py
@@ -1,2 +1,0 @@
-from .grpc_server import DSInferenceGRPCServer
-from .model import DSInferenceModel
