--- bloom-inference-server/ds_zero/__init__.py
@@ -1,1 +1,0 @@
-from .model import DSZeROModel
