--- bloom-inference-server/hf_accelerate/__init__.py
@@ -1,1 +1,0 @@
-from .model import HFAccelerateModel
