+++ bloom-inference-server/utils/__init__.py
@@ -1,5 +1,4 @@
 from .constants import BENCHMARK, CLI, DS_INFERENCE, DS_ZERO, HF_ACCELERATE, SERVER
-from .model import Model, get_downloaded_model_path
 from .requests import (
     GenerateRequest,
     GenerateResponse,
